requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2

# AI Services
deepgram-sdk==3.2.7
openai==1.3.7
elevenlabs==0.2.27

# AWS Services
boto3==1.34.0
botocore==1.34.0

# Telnyx
telnyx==2.0.0

# Utilities
python-dotenv==1.0.0
asyncio-mqtt==0.11.1

# Database (optional - for production)
supabase==2.17.0
psycopg2-binary==2.9.9

# Monitoring/Logging
structlog==23.2.0